+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add breadcrumbs in files UI.
 - Add schema and reset web UI buttons.
+- Allow restrict web UI access by IP addresses.
 
 ## [0.5.0] - 2023-12-31